Refactor image embedding to use float[] and pgvector

+++ src/main/java/com/ds/project/business/v1/services/CbirService.java
@@ -86,7 +86,7 @@ public class CbirService {
         }
     }
 
-    public void pushFeatureToFlask(String id, String productVariantId, String imagePath, Double[] features) {
+    public void pushFeatureToFlask(String id, String productVariantId, String imagePath, float[] features) {
         String url = apiUrl + "/add-features";
         try {
             HttpHeaders headers = new HttpHeaders();
